make media be an effect
