+++ src/frontend/components/Commands/TextCommandEditor.vue
@@ -48,6 +48,8 @@
                   v-model="item.effects"
                   :item-variables="item.variables"
                   :global-variables="globalVariables"
+                  :base-volume="baseVolume"
+                  :widget-url="widgetUrl"
                 />
               </td>
             </tr>
@@ -187,6 +189,14 @@ export default defineComponent({
     channelPointsCustomRewards: {
       type: Object as PropType<Record<string, string[]>>,
       required: true,
+    },
+    baseVolume: {
+      default: 100,
+    },
+    widgetUrl: {
+      type: String,
+      required: false,
+      default: "",
     },
   },
   emits: ["update:modelValue", "cancel"],
