+++ src/frontend/views/CommandsView.vue
@@ -96,7 +96,6 @@ const channelPointsCustomRewards = ref<Record<string, string[]>>({})
 let ws: WsClient | null = null
 const possibleActions: CommandAction[] = [
   CommandAction.TEXT,
-  CommandAction.MEDIA,
   CommandAction.MEDIA_VOLUME,
   CommandAction.COUNTDOWN,
   CommandAction.MADOCHAN_CREATEWORD,
